backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
spacy==3.7.2
textblob==0.17.1
faiss-cpu==1.7.4
psycopg2-binary==2.9.9
scikit-learn==1.4.0
numpy==1.26.2
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
cloudinary==1.36.0
pillow==10.1.0
python-dotenv==1.0.0